pyproject.toml
[project]
name = "geneml"
version = "0.1.0"
description = "Gene Annotation Across Diverse Fungal Species Using Deep Learning"
readme = "README.md"
authors = [
    { name = "Lawrence Hon", email = "lhon@hexagonbio.com" },
    { name = "Lisa Vader", email = "lisvad@biosustain.dtu.dk" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Development Status :: 3 - Alpha"
]
requires-python = ">=3.9"
dependencies = [
    "biopython>=1.78",
    "gitpython>=3.1.45",
    "helperlibs>=0.2.2",
    "numba>=0.53.1",
    "numpy>=1.21.0",
    "psutil>=5.9.0",
    "silence_tensorflow>=0.2.0",
    "tensorflow>=2.17.0;sys_platform != 'linux'",
    # to avoid tensorflow warning messages: https://github.com/tensorflow/tensorflow/issues/54499#issuecomment-1050103601
    "tensorflow-cpu>=2.17.0;sys_platform == 'linux'",
    "tqdm>=4.62.0",
]
license = "GPL-3.0-only"
license-files = [
    "LICENSE",
]

[project.urls]
Repository = "https://github.com/hexagonbio/geneML"

[project.scripts]
geneml = "geneml.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E501"]
